+++ src/main/java/com/api/trip/common/security/config/SecurityConfig.java
@@ -1,5 +1,7 @@
-package com.api.trip.common.security;
+package com.api.trip.common.security.config;
 
+import com.api.trip.common.security.jwt.JwtTokenFilter;
+import com.api.trip.common.security.jwt.JwtTokenProvider;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
